+++ enrolment/tests/test_forms.py
@@ -40,63 +40,36 @@ def test_verification_code_with_email():
     assert isinstance(form.fields['email'], CharField)
 
 
-@mock.patch.object(helpers, 'get_company_profile', return_value={
-    'company_status': 'active',
-})
-def test_companies_house_search_company_number_incomplete_data(client):
-    expected = 'Please contact support to register a Royal Charter Company.'
-    form = forms.CompaniesHouseSearch(
-        data={'company_name': 'Thing', 'company_number': 'RC232323'},
-        session=client.session
-    )
-
-    assert form.is_valid() is False
-    assert form.errors['company_name'] == [expected]
-
-
 def test_companies_house_search_company_number_empty(client):
-    form = forms.CompaniesHouseSearch(
-        data={'company_name': 'Thing'},
-        session=client.session
-    )
+    form = forms.CompaniesHouseCompanySearch(data={'company_name': 'Thing'})
 
     assert form.is_valid() is False
     assert form.errors['company_name'] == [form.MESSAGE_COMPANY_NOT_FOUND]
 
 
 def test_companies_house_search_company_name_empty(client):
-    form = forms.CompaniesHouseSearch(
-        data={},
-        session=client.session
-    )
+    form = forms.CompaniesHouseCompanySearch(data={})
 
     assert form.is_valid() is False
     assert form.errors['company_name'] == ['This field is required.']
 
 
-@mock.patch.object(helpers, 'get_company_profile', return_value={
-    'company_status': 'dissolved',
-})
-def test_companies_house_search_company_dissolved(client):
-    form = forms.CompaniesHouseSearch(
-        data={'company_name': 'Thing', 'company_number': '23232323'},
-        session=client.session
-    )
-
-    assert form.is_valid() is False
-    assert form.errors['company_name'] == [form.MESSAGE_COMPANY_NOT_ACTIVE]
-
-
-@mock.patch.object(helpers, 'get_company_profile', return_value={
-    'company_status': 'active',
-})
-def test_companies_house_search_company_active(client):
-    form = forms.CompaniesHouseSearch(
-        data={'company_name': 'Thing', 'company_number': '23232323'},
-        session=client.session
-    )
-
-    assert form.is_valid() is True
+@pytest.mark.parametrize('status,expected', (
+    ('active', True),
+    ('dissolved', False),
+    ('liquidation', False),
+    ('receivership', False),
+    ('administration', False),
+    ('voluntary-arrangement', True),
+    ('converted-closed', False),
+    ('insolvency-proceedings', False),
+))
+def test_companies_house_search_company_status(client, status, expected):
+    with mock.patch.object(helpers, 'get_companies_house_profile', return_value={'company_status': status}):
+        form = forms.CompaniesHouseCompanySearch(data={'company_name': 'Thing', 'company_number': '23232323'})
+        assert form.is_valid() is expected
+        if expected is False:
+            assert form.errors['company_name'] == [form.MESSAGE_COMPANY_NOT_ACTIVE]
 
 
 @pytest.mark.parametrize('address,expected', (